Specify source file encoding as UTF-8
This setting specifies how build plugins and IDEs should treat files
in the project. If it is not specified, then the editor default for
file creation and editing, and the build system default for analyzing
content, defaults to the default encoding of the current system. The
setting is necessary to ensure that the project behaves the same way
across all operating systems and locales.

Change-Id: I5a0b78dce03b7503451064bda7453a7def156e88